OFFOF should return a short and not a long

+++ bbs/xinit.cpp
@@ -123,7 +123,7 @@ unsigned char WApplication::stryn2tf(const char *s) {
 
 // end callback addition
 
-#define OFFOF(x) ( reinterpret_cast<long>( &session()->user()->data.x ) - reinterpret_cast<long>( &session()->user()->data ) )
+#define OFFOF(x) static_cast<int16_t>(reinterpret_cast<long>(&session()->user()->data.x) - reinterpret_cast<long>(&session()->user()->data))
 
 // Reads WWIV.INI info from [WWIV] subsection, overrides some config.dat
 // settings (as appropriate), updates config.dat with those values. Also
